Move some classes to seperate file
