+++ src/Engine/ProtoCore/DSASM/Heap.cs
@@ -190,35 +190,6 @@ namespace ProtoCore.DSASM
         }
     }
 
-    public class DSArray : HeapElement
-    {
-        private Dictionary<StackValue, StackValue> Dict;
-        public DSArray(int size)
-            : base(size)
-        {
-            Dict = new Dictionary<StackValue, StackValue>();
-            MetaData = new MetaData { type = (int)PrimitiveType.kTypeArray };
-        }
-    }
-
-    public class DSObject : HeapElement
-    {
-        public DSObject(int size)
-            : base(size)
-        {
-            MetaData = new MetaData { type = (int)PrimitiveType.kTypePointer };
-        }
-    }
-
-    public class DSString : HeapElement
-    {
-        public DSString(int size)
-            : base(size)
-        {
-            MetaData = new MetaData { type = (int)PrimitiveType.kTypeString };
-        }
-    }
-
     public class StackValueComparer : IEqualityComparer<StackValue>
     {
         private RuntimeCore runtimeCore;
@@ -322,6 +293,11 @@ namespace ProtoCore.DSASM
 
         public Heap()
         {
+        }
+
+        public T Cast<T>(HeapElement hp) where T : HeapElement
+        {
+            return hp as T; 
         }
 
         /// <summary>
