+++ src/Engine/ProtoCore/ProtoCore.csproj
@@ -96,6 +96,7 @@
     </Compile>
     <Compile Include="AssociativeGraph.cs" />
     <Compile Include="ASTCompilerUtils.cs" />
+    <Compile Include="DSASM\HeapObjects.cs" />
     <Compile Include="SyntaxAnalysis\AstVisitor.cs" />
     <Compile Include="AttributeEntry.cs" />
     <Compile Include="BuildStatus.cs" />
@@ -236,10 +237,9 @@
       <LastGenOutput>Resources.en-US.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
-
   <Target Name="AfterBuild">
     <ItemGroup>
-		<ExtendedLibraries Include="$(ProjectDir)ExtendedLibraries\*.ds" />
+      <ExtendedLibraries Include="$(ProjectDir)ExtendedLibraries\*.ds" />
     </ItemGroup>
     <Copy SourceFiles="@(ExtendedLibraries)" DestinationFolder="$(OutputPath)" />
   </Target>
